src/my_kfp_project_example/pyproject.toml
[project]
name = "my-kfp-project-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docker>=7.1.0",
    "dotenv>=0.9.9",
    "kfp>=2.13.0",
    "kfp-kubernetes>=1.5.0",
]
